Move get_high_water_offsets out of client
